requirements.txt
pyTelegramBotAPI
pyyaml
requests
yandex-weather-api
aiohttp
python-telegram-bot